Handling default value for strides in avg_pool2d and max_pool2d (#780)
* Fix type inference in non-zero and handling default strides in avg_pool for pytorch converter

* Added tests and allowed default strides for max pool 2d

* Fix tests for non-zero layer in MIL

+++ coremltools/converters/mil/mil/ops/tests/test_tensor_operation.py
@@ -603,7 +603,7 @@ class TestNonZero:
         def build(x):
             return [mb.non_zero(x=x)]
 
-        expected_output_types = [(UNK_SYM, 2, types.fp32)]
+        expected_output_types = [(UNK_SYM, 2, types.int)]
         expected_outputs = [np.array(np.transpose(np.nonzero(x_val)))]
 
         run_compare_builder(
